post with detail done

+++ src/app/admin/admin-dashboard/admin-dashboard.component.ts
@@ -1,5 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { ApiHandlerService } from 'src/app/api-handler.service';
+import { environment } from "../../../environments/environment";
+import * as alertify from 'alertify.js';
 
 @Component({
   selector: 'app-admin-dashboard',
@@ -8,6 +10,7 @@ import { ApiHandlerService } from 'src/app/api-handler.service';
 })
 export class AdminDashboardComponent implements OnInit {
   posts: any;
+  imgBase = environment.baseUrl;
   constructor(private api: ApiHandlerService) { }
 
   ngOnInit(): void {
@@ -15,15 +18,24 @@ export class AdminDashboardComponent implements OnInit {
   }
   getPosts() {
     this.api.get('/posts').subscribe((data: any) => {
-      console.log(data);
+      console.log('All Posts:', data.posts);
       this.posts = data.posts;
     });
   }
   deletePost(id) {
     debugger;
-    this.api.delete('/posts/' + id).subscribe(data => {
-      console.log("post deleted", data);
-      alert('post deleted');
-    })
+    if (confirm("Are you sure to delete this post?")) {
+
+      this.api.delete('/posts/' + id).subscribe(data => {
+        console.log("post deleted", data);
+        // alert('post deleted');
+        alertify.success("Post Deleted Successfilly");
+        this.posts = this.posts.filter(function (obj) {
+          return obj._id !== id;
+        });
+      })
+    }
+
   }
+
 }
